+++ src/misc/changelog.ts
@@ -2,6 +2,10 @@
 // DO NOT MODIFY DIRECTLY. Execute `npm run build:changelog` instead.
 
 export const changelog = [
+	{ title: "Version 10", content: [
+		"Balance",
+		"• When a sheep destroys their own farms, gold coins are left behind equal to half the bounty of the farm",
+	].join( "\n" ) },
 	{ title: "Version 9", content: [
 		"Balance",
 		"• Wisp",
@@ -41,6 +45,7 @@ export const changelog = [
 		"  - Right-clicking 12 farms or more away, ordered units are teleported",
 		"• Can no longer give resources via Allies menu",
 		"• Added -lumber (-l) command",
+		"• Added w3mmd variables: farms built/destroyed, sheep kills/saves/deaths, levels, specializations, gold",
 		"Bug fixes",
 		"• Hidden Saving Farms have correct scaling (1.15 -> 1)",
 		"• Fixed Discord link",